requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
supabase>=2.5.0
requests>=2.32.0
python-dotenv>=1.0.1
trafilatura>=1.9.0
lxml_html_clean>=0.1.1
pdfplumber>=0.11.0
python-docx>=1.2.0
feedparser>=6.0.11
google-api-python-client>=2.188.0
google-auth-oauthlib>=1.2.4
